Move KeyPair generation to a single file
